requirements.txt
Django==1.6.1
Jinja2==2.7.1
MarkupSafe==0.18
South==0.8.4
argparse==1.2.1
distribute==0.6.24
dj-database-url==0.2.2
django-bootstrap-form==3.1
django-extra-views==0.6.3
ffs==0.0.7.4
gunicorn==18.0
letter==0.4.2
six==1.4.1
wsgiref==0.1.2
psycopg2
dj_static
django-reversion==1.8.0
